sales_feedback/requirements.txt
# 営業フィードバックエージェント 依存関係

# AI分析
google-generativeai==0.8.3

# 音声文字起こし
assemblyai==0.33.0

# Google連携
google-api-python-client==2.155.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Notion連携
notion-client==2.2.1

# Slack連携
slack-sdk==3.33.1

# ユーティリティ
requests==2.32.3
python-dotenv==1.0.1
